requirements.txt
horovod==0.23.0
joblib==1.0.1
numpy==1.20.3
pandas==1.3.3
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.7.1
spacy==2.3.7
torch==1.8.2+cu111
tqdm==4.62.2
transformers==4.12.2
